requirements.txt
aiomysql==0.1.1
aiosqlite==0.17.0
amqp==5.1.1
asgiref==3.6.0
attrs==22.2.0
billiard==3.6.4.0
CacheControl==0.12.11
cachetools==5.3.0
celery==5.2.3
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==39.0.1
Deprecated==1.2.13
ecdsa==0.18.0
exceptiongroup==1.1.0
fastapi==0.68.2
firebase-admin==6.1.0
flower==1.0.0
gcloud==0.18.3
google-api-core==2.11.0
google-api-python-client==2.77.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.9.1
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
h11==0.14.0
httplib2==0.21.0
humanize==4.6.0
idna==3.4
iniconfig==2.0.0
iso8601==1.1.0
jwcrypto==1.4.2
kombu==5.2.4
msgpack==1.0.4
oauth2client==4.1.3
packaging==23.0
passlib==1.7.4
pluggy==1.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
proto-plus==1.22.2
protobuf==4.22.0rc2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.5
PyJWT==2.6.0
PyMySQL==1.0.2
pyparsing==3.0.9
pypika-tortoise==0.1.6
Pyrebase4==4.6.0
pytest==7.2.1
python-jose==3.3.0
python-jwt==4.0.0
python-multipart==0.0.5
pytz==2022.7.1
redis==3.5.3
requests==2.28.2
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
starlette==0.14.2
tomli==2.0.1
tornado==6.2
tortoise-orm==0.19.2
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
uvicorn==0.15.0
vine==5.0.0
wcwidth==0.2.6
wrapt==1.14.1
